+++ packages/service-core-tests/src/test-utils/test-utils-index.ts
@@ -1,4 +1,4 @@
 export * from './bucket-validation.js';
 export * from './general-utils.js';
-export * from './metrics-utils.js';
+export * from './MetricsHelper.js';
 export * from './stream_utils.js';